+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.2.1-m2.2
+- sprzętowo potwierdzone AVRCP Play/Pause/Next/Previous
+- dwukierunkowa synchronizacja głośności z telefonem
+- metadata artist/title oraz peer/device info
+- stabilna praca Bluetooth A2DP, Wi-Fi, WWW i OTA równolegle
+- TFT z metadata i częściowym odświeżaniem bez migotania
+
 ## 0.1.1-m1
 - pierwsze potwierdzone OTA przez WWW
 - marker `OTA TEST OK`